Refactor: encapsulate `RegionProperties`
