+++ src/main/java/com/cheffi/region/controller/RegionController.java
@@ -1,4 +1,4 @@
-package com.cheffi.review.controller;
+package com.cheffi.region.controller;
 
 import java.util.List;
 
@@ -6,8 +6,9 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
-import com.cheffi.common.config.data.RegionProperties;
+import com.cheffi.region.service.RegionDto;
 import com.cheffi.common.response.ApiResponse;
+import com.cheffi.region.service.RegionService;
 
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.tags.Tag;
@@ -18,14 +19,13 @@ import lombok.RequiredArgsConstructor;
 @RequiredArgsConstructor
 public class RegionController {
 
-	private final RegionProperties properties;
+	private final RegionService regionService;
 
 	@Tag(name = "regionInfo")
 	@Operation(summary = "시, 군/구 정보 조회 API, 세종시는 군/구대신 읍/면 까지 조회합니다.")
 	@GetMapping
-	public ApiResponse<List<RegionProperties.Region>> getAllRegionInfo() {
-
-		return ApiResponse.success(properties.getRegions());
+	public ApiResponse<List<RegionDto>> getAllRegionInfo() {
+		return ApiResponse.success(regionService.getRegion());
 	}
